Simplify problem module import
The previous commit didn't consider the case where a problem module
is found on Python's path elsewhere than the problems directory or the
current directory.

Instead of checking if multiple problem modules are found; simply print
the origin of the module that is imported.

In the future (Python >= 3.7) we could use ModuleNotFoundError instead.